loongarch: Fix up loongarch_atomic_assign_expand_fenv
Commit message shamelessly copied from 1777beb6b129 by jakub:

This function, because it is sometimes called even outside of function
bodies, uses create_tmp_var_raw rather than create_tmp_var.  But in order
for that to work, when first referenced, the VAR_DECLs need to appear in a
TARGET_EXPR so that during gimplification the var gets the right
DECL_CONTEXT and is added to local decls.

gcc/

	* config/loongarch/loongarch-builtins.c: Use TARGET_EXPR instead
	  of MODIFY_EXPR.